Multiple Layer GRU
Accuracy was almost reached 100% on the training set and was approx. 85% on validation set. But the loss on the validation set started increasing after 0.5 reaching approx. max value of 0.8. At the same time the loss on the training set showed a decreasing trend reaching a minimum value in the range: 0 to 0.1. This is a clear overfitting scenario.