Cosine similarity, word analogy, and Debiasing
This commit comprises of the following sections:
1 - Cosine similarity: To measure the similarity between two words, we need a way to measure the degree of similarity between two embedding vectors for the two words. 

Cosine similarity is a good way to compare the similarity between pairs of word vectors.
Note that L2 (Euclidean) distance also works.
For NLP applications, using a pre-trained set of word vectors is often a good way to get started.

2 - Word analogy task: In the word analogy task, we complete the sentence:
"a is to b as c is to __".

An example is: 'man is to woman as king is to queen' .

We are trying to find a word d, such that the associated word vectors ea, eb, ec, ed are related in the following manner: eb−ea≈ed−ec
 
We will measure the similarity between eb−ea and ed−ec using cosine similarity.

3 - Debiasing word vectors
a. Neutralize bias for non-gender specific words
b. Equalization algorithm for gender-specific words

These debiasing algorithms are very helpful for reducing bias, but are not perfect and do not eliminate all traces of bias. For example, one weakness of this implementation was that the bias direction g was defined using only the pair of words woman and man. As discussed earlier, if g were defined by computing g1=ewoman−eman; g2=emother−efather; g3=egirl−eboy; and so on and averaging over them, you would obtain a better estimate of the "gender" dimension in the 50 dimensional word embedding space. Feel free to play with such variants as well.